美食分享平台/requirements.txt
Flask==2.3.3
pymongo==4.5.0
flask-pymongo==2.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
flask-login==0.6.2
flask-wtf==1.1.1
Werkzeug==2.3.7
Pillow==10.0.0